Docs: 세팅페이지 컴포넌트 파일 정리
