+++ CaffeineDrop/components/settingPage/CheckIcon.js
@@ -1,21 +1,20 @@
 import React from "react";
 import Svg, { Rect, Path } from "react-native-svg";
 import styled from "styled-components/native";
-import { TouchableOpacity } from "react-native";
 import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
-const CheckedIconWrapper = styled.View`
+const CheckIconWrapper = styled.View`
   width: ${responsiveWidth(20)}px;
   height: ${responsiveHeight(20)}px;
   flex-shrink: 0;
 `;
 
-const CheckedIcon = () => (
-  <CheckedIconWrapper>
+const CheckIcon = () => (
+  <CheckIconWrapper>
     <Svg
       xmlns="http://www.w3.org/2000/svg"
       width={`${responsiveWidth(20)}px`}
@@ -27,18 +26,17 @@ const CheckedIcon = () => (
         width={`${responsiveWidth(20)}px`}
         height={`${responsiveHeight(20)}px`}
         rx="4"
-        fill="#756555"
+        fill="#F1F1F1"
       />
-      <Path d="M5.5 10L8.40559 13L14 6" fill="#756555" />
       <Path
         d="M5.5 10L8.40559 13L14 6"
-        stroke="#FAFAFA"
+        stroke="#756555"
         strokeWidth="1.5"
         strokeLinecap="round"
         strokeLinejoin="round"
       />
     </Svg>
-  </CheckedIconWrapper>
+  </CheckIconWrapper>
 );
 
-export default CheckedIcon;
+export default CheckIcon;
