+++ CaffeineDrop/components/settingPage/KaKaoIcon.js
@@ -4,7 +4,7 @@ import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 export default function KaKaoIcon(props) {
   return (
