+++ CaffeineDrop/components/settingPage/NextButton.js
@@ -4,7 +4,7 @@ import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 const NextButton = (props) => (
   <Svg
